refactor(core): gateway 与 query 域迁移到 src/domains 并接入门面
