+++ dev-docs/engineering/event-catalog.md
@@ -9,5 +9,5 @@
 | `user` | noj-core/src/domains/community/services/notifications.ts<br>noj-core/src/domains/messaging/routes/conversations.ts<br>noj-core/src/domains/messaging/services/messages.ts<br>noj-core/src/routes/sse.ts |
 | `contestRanking` | noj-core/src/domains/submission/services/submissions/submissions-result.ts<br>noj-core/src/routes/sse.ts |
 | `contestSubmission` | noj-core/src/domains/submission/services/submissions/artifact-submissions.ts<br>noj-core/src/domains/submission/services/submissions/submissions-crud.ts<br>noj-core/src/routes/sse.ts |
-| `stats` | noj-core/src/routes/sse.ts<br>noj-core/src/services/stats-cache.ts |
+| `stats` | noj-core/src/domains/query/services/stats-cache.ts<br>noj-core/src/routes/sse.ts |
 | `announcements` | noj-core/src/domains/system/routes/announcements.ts<br>noj-core/src/domains/system/services/announcements.ts |
