+++ noj-core/src/app.ts
@@ -2,7 +2,7 @@ import { Hono } from "hono";
 import type { Context, Next } from "hono";
 import { cors } from "hono/cors";
 import health from "./routes/health.ts";
-import stats from "./routes/stats.ts";
+import stats from "./domains/query/routes/stats.ts";
 import auth from "./domains/identity/routes/auth.ts";
 import admin from "./routes/admin/index.ts";
 import adminAnnouncements from "./domains/system/routes/admin-announcements.ts";
@@ -14,11 +14,11 @@ import queue from "./domains/submission/routes/queue.ts";
 import submissions from "./domains/submission/routes/submissions.ts";
 import selfTests from "./domains/submission/routes/self-tests.ts";
 import users from "./domains/identity/routes/users.ts";
-import rankings from "./routes/rankings.ts";
+import rankings from "./domains/query/routes/rankings.ts";
 import conversations from "./domains/messaging/routes/conversations.ts";
 import community from "./domains/community/routes/community.ts";
 import communityAdmin from "./domains/community/routes/community-admin.ts";
-import search from "./routes/search.ts";
+import search from "./domains/query/routes/search.ts";
 import contests from "./domains/contest/routes/contests.ts";
 import trainings from "./domains/catalog/routes/trainings.ts";
 import announcements from "./domains/system/routes/announcements.ts";
