+++ noj-core/src/domains/gateway/routes/admin-llm.ts
@@ -1,6 +1,6 @@
 import { Hono } from "hono";
-import type { AuthEnv } from "../../middleware/auth.ts";
-import { parseJsonBody } from "../../lib/request.ts";
+import type { AuthEnv } from "../../../middleware/auth.ts";
+import { parseJsonBody } from "../../../lib/request.ts";
 import {
   createLlmProvider,
   listLlmProviders,
@@ -10,7 +10,7 @@ import {
   queryLlmUsage,
   updateLlmProvider,
   upsertLlmQuota,
-} from "../../services/llm.ts";
+} from "../services/llm.ts";
 
 /**
  * 管理端 LLM Gateway 路由（挂载前缀 /api/v1/admin）。
