+++ noj-core/src/domains/query/index.ts
@@ -1,10 +1,4 @@
-export {
-  applyNewResult,
-  getCachedTodayStats,
-  getCachedTotalStats,
-} from "../../services/stats-cache.ts";
-export {
-  getGlobalRankings,
-  getMyRanking,
-  refreshRankingsView,
-} from "../../services/rankings.ts";
+export * from "./services/search.ts";
+export * from "./services/rankings.ts";
+export * from "./services/stats-cache.ts";
+export * from "./services/dashboard.ts";
