+++ noj-core/src/domains/query/routes/admin-dashboard.ts
@@ -1,6 +1,6 @@
 import { Hono } from "hono";
-import type { AuthEnv } from "../../middleware/auth.ts";
-import { getDashboardStats } from "../../services/dashboard.ts";
+import type { AuthEnv } from "../../../middleware/auth.ts";
+import { getDashboardStats } from "../services/dashboard.ts";
 
 /**
  * 管理端仪表盘路由（挂载前缀 /api/v1/admin，见 admin/index.ts）。
