+++ noj-core/src/domains/query/routes/stats.ts
@@ -1,12 +1,12 @@
 import { Hono } from "hono";
 import { and, count, eq, sql } from "drizzle-orm";
-import { getDb } from "../db/connection.ts";
+import { getDb } from "../../../db/connection.ts";
 import {
   evaluationResults,
   problems,
   submissions,
   users,
-} from "../db/schema.ts";
+} from "../../../db/schema.ts";
 
 const stats = new Hono();
