+++ noj-core/src/domains/query/services/search.ts
@@ -12,8 +12,8 @@
  */
 
 import { sql } from "drizzle-orm";
-import { getDb } from "../db/connection.ts";
-import { ForbiddenError } from "../lib/errors.ts";
+import { getDb } from "../../../db/connection.ts";
+import { ForbiddenError } from "../../../lib/errors.ts";
 
 /**
  * 转义 LIKE 模式中的元字符（reviewer issue 2）。
