+++ noj-core/src/domains/query/services/stats-cache.ts
@@ -1,9 +1,9 @@
 import { and, eq, gte, sql } from "drizzle-orm";
-import { todayUtc } from "../lib/dates.ts";
-import { getDb } from "../db/connection.ts";
-import { evaluationResults, submissions } from "../db/schema.ts";
-import { Channels, publishSseEvent } from "../lib/event-bus.ts";
-import { FULL_SCORE } from "../lib/constants.ts";
+import { todayUtc } from "../../../lib/dates.ts";
+import { getDb } from "../../../db/connection.ts";
+import { evaluationResults, submissions } from "../../../db/schema.ts";
+import { Channels, publishSseEvent } from "../../../lib/event-bus.ts";
+import { FULL_SCORE } from "../../../lib/constants.ts";
 
 export interface StatsSnapshot {
   total: number;
