+++ noj-core/tests/routes/rankings.test.ts
@@ -1,6 +1,6 @@
 import { assertEquals, assertExists } from "jsr:@std/assert@^1";
 import { Hono } from "hono";
-import rankings from "../../src/routes/rankings.ts";
+import rankings from "../../src/domains/query/routes/rankings.ts";
 import { AppError } from "../../src/lib/errors.ts";
 import { getDb, resetDbForTest } from "../../src/db/connection.ts";
 import { checkIns, users } from "../../src/db/schema.ts";
